Add routing to 404-page
