+++ frontend/app/routes/not-found.tsx
@@ -1,0 +1,28 @@
+import { useTranslation } from "react-i18next";
+import NavigateButton from "~/shared/components/NavigateButton";
+
+export default function NotFound() {
+  const { t } = useTranslation();
+
+  return (
+    <main>
+      <title>{`${t("nav.home")} | VIERNULVIER`}</title>
+        <div className="mx-auto max-w-5xl py-12 text-center md:py-24">
+            <h1 className="mb-6 font-serif text-5xl italic md:mb-12 md:text-8xl">
+                {"VIERNULVIER - " + t("notFound.title")}
+            </h1>
+            <p className="mb-6 text-lg md:mb-12 md:text-2xl">
+                {t("notFound.description")}
+            </p>
+            <div className="flex flex-col items-center justify-center gap-4 md:flex-row">
+                <NavigateButton location="archive" name={t("home.buttons.explore")} />
+                <NavigateButton
+                    location="history"
+                    name={t("home.buttons.history")}
+                    variant="secondary"
+                />
+            </div>
+        </div>           
+    </main>
+  );
+}
