+++ frontend/public/locales/en/translation.json
@@ -106,5 +106,12 @@
         "generic": "We could not sign you in right now. Please try again."
       }
     }
+  },
+  "notFound" : {
+    "title": "Page not found",
+    "description": "The page you are looking for does not exist. It might have been removed, had its name changed, or is temporarily unavailable.",
+    "GoToHome": "Go to Home",
+    "GoToArchive": "Go to Archive",
+    "GoToHistory": "Go to History"
   }
 }
