+++ frontend/public/locales/nl/translation.json
@@ -106,5 +106,12 @@
         "generic": "Aanmelden lukt momenteel niet. Probeer het opnieuw."
       }
     }
+  },
+  "notFound" : {
+    "title": "Pagina niet gevonden",
+    "description": "De pagina die u zoekt bestaat niet. Deze kan zijn verwijderd, de naam gewijzigd of tijdelijk niet beschikbaar.",
+    "GoToHome": "Ga naar Home",
+    "GoToArchive": "Ga naar Archive",
+    "GoToHistory": "Ga naar History"
   }
 }
